admin page moved to pages from admin dir
